lifeguard-ai/backend/requirements.txt
fastapi
uvicorn[standard]
openai
pinecone
sentence-transformers
pydantic
python-dotenv
httpx
pandas
numpy
